Export subpath helpers
